pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bytesnip"
version = "0.1.0"
description = "A fast, keyboard-first code snippet manager for macOS and Linux"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "pygments>=2.18",
    "pynput>=1.7",
    "sqlalchemy>=2.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Cocoa>=10.0",
    "pyobjc-framework-Quartz>=10.0",
    "pyobjc-framework-ApplicationServices>=10.0",
]
linux = [
    "python-xlib>=0.33",
]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "ruff>=0.4",
    "mypy>=1.10",
    "pyinstaller>=6.0",
]

[project.scripts]
bytesnip = "snipapp.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/snipapp"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
